+++ src/MimesisPlayerEnhancement/Config/ModConfig.cs
@@ -177,8 +177,8 @@ namespace MimesisPlayerEnhancement
         public static MelonPreferences_Entry<bool> EnableExtendedSaveSlots { get; internal set; } = null!;
         public static MelonPreferences_Entry<bool> EnableExtendedSpectatorPlayerList { get; internal set; } = null!;
         public static MelonPreferences_Entry<bool> EnableExtendedInGameMenuPlayerList { get; internal set; } = null!;
-        public static MelonPreferences_Entry<bool> EnableWorldHealthBars { get; internal set; } = null!;
-        public static MelonPreferences_Entry<float> WorldHealthBarDurationSeconds { get; internal set; } = null!;
+        public static MelonPreferences_Entry<bool> EnableWorldHealthGlow { get; internal set; } = null!;
+        public static MelonPreferences_Entry<float> WorldHealthGlowDurationSeconds { get; internal set; } = null!;
         public static MelonPreferences_Entry<bool> EnableFloatingDamageNumbers { get; internal set; } = null!;
         public static MelonPreferences_Entry<float> FloatingDamageDurationSeconds { get; internal set; } = null!;
         public static MelonPreferences_Entry<bool> EnableFloatingDetoxIndicators { get; internal set; } = null!;