+++ thunderstore/CHANGELOG.md
@@ -1,10 +1,12 @@
 Hint: this mod is still in Alpha - use at your own risk! These updates can break things that worked before. Still needs extensive testing. In case of any errors please create an issue: https://github.com/Kandru/mimesis-player-enhancements/issues
 
 ## 26.7.11
+- feature: replace health bar with health glow
 - improvement: add another button to the items overlay to allow giving multiple items without re-opening the items overlay
 - fix: double config sync - global config only will be saved when closing the game properly
 - fix: save/load all data when saving or loading mod-data
 - fix: steamid / voiceid was connection was not saved properly (so voice data could not be loaded)
+- fix: minimap had some smaller issues
 - draft: added privacy settings to disallow the game to send data to anything else then steam
 - code cleanup (may leads to issues further down the road, we will see)